Force noEmit compiler option to be false
When a developer makes a mistake and has noEmit with a value of true in their tsconfig.json, this plugin errors out with "Typescript compilation failed" message.
The message is wrong, as compilation actually did succeed. The real problem is that there  simply were no files written to disk and as such there are no files for serverless to deploy.

To prevent developers from making this mistake, its value can be forced by the plugin.